Show more items in tradehistory
Calling `/history` now shows 5 items as a default.
This can be adjusted to your liking.
The reason to show more items than 1 is because there are times when
a limit order is placed that it is only partially fulfilled.
This shows up as multiple orders at the limit, saving pressing
next a lot of times.